gkdi is now fixed, dce samples added
