+++ aiosmb/_version.py
@@ -1,5 +1,5 @@
 
-__version__ = "0.4.7"
+__version__ = "0.4.8"
 __banner__ = \
 """
 # aiosmb %s
